+++ driver-core/src/test/java/com/datastax/driver/core/HostAssert.java
@@ -53,13 +53,6 @@ public class HostAssert extends AbstractAssert<HostAssert, Host> {
         return this;
     }
 
-    public HostAssert isNotReconnectingFromSuspected() {
-        assertThat(actual.getInitialReconnectionAttemptFuture() != null && !actual.getInitialReconnectionAttemptFuture().isDone())
-            .isFalse();
-        return this;
-
-    }
-
     public HostAssert comesUpWithin(long duration, TimeUnit unit) {
         final CountDownLatch upSignal = new CountDownLatch(1);
         StateListener upListener = new StateListenerBase() {