+++ driver-core/CHANGELOG.rst
@@ -40,6 +40,7 @@ CHANGELOG
   schema-altering statement (JAVA-669)
 - [improvement] Make connection and pool creation fully async (JAVA-692)
 - [improvement] Optimize connection use after reconnection (JAVA-505)
+- [improvement] Remove "suspected" mechanism (JAVA-617)
 
 Merged from 2.0.9_fixes branch: